apps/diarizer/requirements.txt
pyannote.audio==3.3.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
torch>=2.0.0
torchaudio>=2.0.0
